Placed the control creation code in another file.
